+++ include/utils/lzw.hpp
@@ -352,7 +352,7 @@ namespace lzw
             bool read(CodeType& k)
             {
                 // ready-made bit masks
-                static const std::array<unsigned long int, 9> masks { {0x00, 0x01, 0x03, 0x07, 0x0F, 0x1F, 0x3F, 0x7F, 0xFF} };
+                static const std::array<size_t, 9> masks { {0x00, 0x01, 0x03, 0x07, 0x0F, 0x1F, 0x3F, 0x7F, 0xFF} };
                 std::size_t remaining_bits {bits};
                 std::size_t offset {lo.used};
                 unsigned char temp;
@@ -406,7 +406,7 @@ namespace lzw
         /// @param n    number to be evaluated
         /// @return Number of required bits.
         ///
-        inline std::size_t required_bits(unsigned long int n)
+        inline std::size_t required_bits(size_t n)
         {
             std::size_t r {1};
